feat(sbase): Room database layer with entities, DAOs, and migrations
Add Room v17 database with 27 entities covering locations, sessions,
activities, Wi-Fi, cell data, and tracker state. Includes DAOs,
type converters, migration paths, and schema exports.

+++ sbase/src/main/java/com/adsamcik/tracker/shared/base/database/data/CellSample.kt
@@ -28,10 +28,10 @@ data class CellSample(
 	val timeMs: Long,
 
 	/**
-	 * Cell tower ID (CID).
+	 * Cell tower ID (CID). Long to support 5G NR NCI values.
 	 */
 	@ColumnInfo(name = "cell_id")
-	val cellId: Int,
+	val cellId: Long,
 
 	/**
 	 * Location Area Code (LAC).
